requirements.txt
python-docx>=1.1,<2
Pillow>=10,<12
pystray>=0.19,<1
